LEXER_REFACTOR setup abstract TokenizerService class

+++ Calc/Token.cpp
@@ -3,36 +3,30 @@
 
 namespace KLab
 {
-	Token::Token(const string &lexame, TOKEN_CLASS tokenClass, int line /*= 0*/, int column /*= 0*/)
-	{
+	Token::Token(const string &lexame, TOKEN_CLASS tokenClass, int line /*= 0*/, int column /*= 0*/) {
 		this->lexeme = lexeme;
 		init(tokenClass, line, column);
 	}
 
-	Token::Token(string &&lexame, TOKEN_CLASS tokenClass, int line /*= 0*/, int column /*= 0*/)
-	{
+	Token::Token(string &&lexame, TOKEN_CLASS tokenClass, int line /*= 0*/, int column /*= 0*/) {
 		this->lexeme = std::move(lexeme);
 		init(tokenClass, line, column);
 	}
 
-	void Token::init(TOKEN_CLASS tokenClass, int line, int column)
-	{
+	void Token::init(TOKEN_CLASS tokenClass, int line, int column) {
 		this->column = column;
 		this->line = line;
 		this->tokenClass = tokenClass;
 	}
 
-	Token::~Token()
-	{
+	Token::~Token() {
 	}
 
-	bool Token::operator==(const Token &second)
-	{
+	bool Token::operator==(const Token &second) {
 		return lexeme == second.lexeme && tokenClass == second.tokenClass;
 	}
 
-	string Token::getLexeme() const
-	{
+	string Token::getLexeme() const {
 		return lexeme;
 	}
 
@@ -40,28 +34,23 @@ namespace KLab
 		return (int)lexeme.length();
 	}
 
-	void Token::setLexeme(const string &lexemeArg)
-	{
+	void Token::setLexeme(const string &lexemeArg) {
 		lexeme = lexemeArg;
 	}
 
-	TOKEN_CLASS Token::getTokenClass() const
-	{
+	TOKEN_CLASS Token::getTokenClass() const {
 		return tokenClass;
 	}
 
-	void Token::setTokenClass(const TOKEN_CLASS &typeArg)
-	{
+	void Token::setTokenClass(const TOKEN_CLASS &typeArg) {
 		tokenClass = typeArg;
 	}
 
-	int Token::getLine() const
-	{
+	int Token::getLine() const {
 		return line;
 	}
 
-	void Token::setLine(int newLine)
-	{
+	void Token::setLine(int newLine) {
 		line = newLine;
 	}
 
@@ -74,16 +63,14 @@ namespace KLab
 	}
 
 	int Token::lastIndexOf(char c) const {
-		return lexeme.find_last_of(c);
+		return (int)lexeme.find_last_of(c);
 	}
 
-	int Token::getColumn()const
-	{
+	int Token::getColumn()const {
 		return column;
 	}
 
-	void Token::setColumn(int pos)
-	{
+	void Token::setColumn(int pos) {
 		column = pos;
 	}
 }
